pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "okta-lifecycle-lab"
version = "0.1.0"
description = "Simulated Okta identity lifecycle automation lab"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Alex Davenport" }]
dependencies = [
    "pydantic>=2.0",
    "typer>=0.9",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.5",
]
mock = [
    "flask>=3.0",
    "flask-cors>=4.0",
]
all = ["okta-lifecycle-lab[dev,mock]"]

[project.scripts]
lifecycle-lab = "engine.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["engine", "mock_scim"]

[tool.mypy]
python_version = "3.11"
strict = true
